+++ catboost/libs/metrics/precision_recall_at_k.h
@@ -2,8 +2,8 @@
 
 #include <util/generic/fwd.h>
 
-double CalcPrecisionAtK(TConstArrayRef<double> approx, TConstArrayRef<float> target, int top, float border);
+double CalcPrecisionAtK(TConstArrayRef<double> approx, TConstArrayRef<float> target, int top, double border);
 
-double CalcRecallAtK(TConstArrayRef<double> approx, TConstArrayRef<float> target, int top, float border);
+double CalcRecallAtK(TConstArrayRef<double> approx, TConstArrayRef<float> target, int top, double border);
 
-double CalcAveragePrecisionK(TConstArrayRef<double> approx, TConstArrayRef<float> target, int top, float border);
+double CalcAveragePrecisionK(TConstArrayRef<double> approx, TConstArrayRef<float> target, int top, double border);
